добавлен фильтр картинок
